evals/scenarios/create-prompt-and-dataset/project/requirements.txt
freeplay-sdk>=0.5.6
openai>=1.0.0
